LPD-60940 Move methods to base to reuse code
